+++ src/cli/export/export.go
@@ -49,7 +49,7 @@ func exportCmd() *cobra.Command {
 		Short: "Export your service data",
 		Long:  `Export the data stored in one of your M365 services.`,
 		RunE:  handleExportCmd,
-		Args:  cobra.NoArgs,
+		Args:  utils.SubcommandsRequiredWithSuggestions,
 	}
 }